Refactor and add tests for data utilities and role checks
Removed unused test files, added new tests for `reset_id` and `read_csv`, improved exception handling, and updated data type constraints. Deprecated and commented out unused range-based approval logic while adding discrete value validation. Expanded role-checking functionality with `check_info_systems_role`.

+++ PledgePoints/csvutils.py
@@ -38,9 +38,7 @@ def read_csv(filename):
     columns = ("ID", "Time", "PointChange", "Pledge", "Brother", "Comment", "Approved")
     if not os.path.exists(filename):
         raise FileNotFoundError("The file {} doesn't exist.".format(filename))
-    dtypes = {columns[0]: "uint32",  # ID should never be negative. The only reason its 32 bits is because I test with a
-              # dataset that is 100,000 rows long and would run into overflow problems otherwise.
-              # If  optimization is needed an uint16 would be just fine.
+    dtypes = {columns[0]: "uint16",  # ID should never be negative.
               columns[2]: "int8",  # This limits point change values to -128 to 127. That shouldn't be an issue, but if
               # you run into overflow problems you can change to int16 or in32.
               columns[3]: "string",
